:sparkles: Allow grouping allomorphs by d/t form
Adds a query option for morphemesByShape: `compareBy`.
This takes a CherokeeOrthography type, like `DT` or `TTH` representing
the d/t and t/th orthographies respectively.
Because the d/t representation of a morpheme is simpler, grouping
instances by it provides a broader grained list of allomorphs for a
particular gloss.