Allow choosing tag set on glossary
Implement tag set picker, similar to the mode picker, that lets users
choose which tag set to see definitions for. The initial selection is
based off of which mode you last chose when viewing a document, but this
tag set selection doesn't propagate back. This is to facilitate checking
out various representations of a morpheme without losing your original
mode selection.