Switch serif font to Alegreya with smallcaps
Instead of using Noto Serif for the body text, this switches us to
Alegreya, which incidentally comes with a smallcaps variant "Alegreya
SC". Since functional morpheme tags are generally typeset with
smallcaps, this is a great addition! The font also just looks better on
the DAILP site, especially when printed.

Another significant gain with Alegreya: the webfont fully supports the
double acute modifier accent, meaning superhigh "v" is no longer screwed
up or rendered with Arial.