Switch to Spectral and Digohweli fonts
Alegreya was looking a bit too small and old timey, so I decided to
switch to Spectral, which also has a nice smallcaps variant. This font
is bigger and fuller to read, while looking more neutral.

I also switched the Cherokee font from Noto Sans Cherokee to Digohweli,
which more closely resembles the traditional syllabary characters and
doesn't rely on Google. As more of an emphatically serif font, the
character details are better preserved and fit more closely with the
other text in Spectral.