Use Charis SIL for body font
Charis SIL is known to have full compatibility with all the symbols that
we want. It has italics since it's the standard font for academic
interlinear glossing. It's clean and clear, probably a more neutral
choice than Spectral. Plus, the digits are more easily distinguishable
in small caps, though a bit tall.