Pin EC2 instance image ID
This should finally fix the deployment issue on AWS.
Before, we were using the "latest Amazon Linux 2" image, which updates
every few months. This commit pins it to the December 2020 image, so
that our MongoDB instances have stable deployment until we *choose* to
upgrade, which will be a semi-manual process.