Render client-only routes on 404 page
TinaCMS allows an authenticated editor to add pages to the site.
Previously, a site rebuild had to occur for those pages to exist to then
be editable. This change leverages the 404 page to render routes that
exist in the database, but haven't been statically rendered yet. This
provides a gradual ramp from dynamic routes to static routes.