Rollback lambda function builds to use MUSL
I thought I could use the GNU libc libraries for lambda functions, but
because we are building with nix, the dynamic library locations won't
work for a lambda. The easiest solution was to continue to use musl
libc, which is entirely statically linked.

Also I added `act` as a local development tool which we can use to run
GitHub actions locally. This works better than cross-compiling lambda
functions from MacOS (which is a pain in the ASS), and as a bonus
ensures that the whole GH action pipeline is working.